requirements.txt
discord.py
aiomysql
cryptography
